w1retap-1.0.1.tar.bz2
sha512sum adc777ffb2b07d67849fc7bf7bb25ac868a75653f49a964724fbb44c8f77e86e6867942654f5009fdd10f3fa9ea9357f4d3d991d133f7432d32f25c0cbd1da05

+++ src/libusblinux300/gethumd.c
@@ -41,7 +41,7 @@
 //----------------------------------------------------------------------
 //  This is the Main routine for debit
 //
-int main(short argc, char **argv)
+int main(int argc, char **argv)
 {
    char msg[200];
    int portnum = 0;
